项目结束
事实证明。Winsock在Win7环境下受到一定的限制。
这最终让我放弃了Winsock这一套API。
Limitations on Raw Sockets
On Windows 7, Windows Vista, Windows XP with Service Pack 2 (SP2), and
Windows XP with Service Pack 3 (SP3), the ability to send traffic over
raw sockets has been restricted in several ways:

TCP data cannot be sent over raw sockets.
UDP datagrams with an invalid source address cannot be sent over raw
sockets. The IP source address for any outgoing UDP datagram must exist
on a network interface or the datagram is dropped. This change was made
to limit the ability of malicious code to create distributed
denial-of-service attacks and limits the ability to send spoofed packets
(TCP/IP packets with a forged source IP address).
A call to the bind function with a raw socket for the IPPROTO_TCP
protocol is not allowed.
Note  The bind function with a raw socket is allowed for other protocols
(IPPROTO_IP, IPPROTO_UDP, or IPPROTO_SCTP, for example).

These above restrictions do not apply to Windows Server 2008 R2, Windows
Server 2008 , Windows Server 2003, or to versions of the operating
system earlier than Windows XP with SP2.